requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==5.0
django-quill-editor==0.1.40
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pycodestyle==2.11.1
sqlparse==0.4.4
tzdata==2023.4
whitenoise==6.6.0
